Conflicts:
	gradoop-giraph/src/main/java/org/gradoop/algorithms/LabelPropagationComputation.java

+++ gradoop-giraph/src/main/java/org/gradoop/algorithms/LabelPropagationComputation.java
@@ -3,20 +3,18 @@ package org.gradoop.algorithms;
 import org.apache.giraph.graph.BasicComputation;
 import org.apache.giraph.graph.Vertex;
 import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.NullWritable;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 
 /**
  * TODO: algorithm description
  */
 public class LabelPropagationComputation extends
-  BasicComputation<LongWritable, LongWritable, NullWritable, LongWritable> {
+  BasicComputation<IntWritable, IntWritable, NullWritable, IntWritable> {
   /**
    * Returns the current new value. This value is based on all incoming
    * messages. Depending on the number of messages sent to the vertex, the
@@ -32,13 +30,13 @@ public class LabelPropagationComputation extends
    * @param messages All incoming messages
    * @return the new Value the vertex will become
    */
-  private long getNewValue(Vertex<LongWritable, LongWritable,
-    NullWritable> vertex, Iterable<LongWritable> messages) {
-    long newValue;
+  private int getNewValue(Vertex<IntWritable, IntWritable, NullWritable> vertex,
+    Iterable<IntWritable> messages) {
+    int newValue;
     //TODO: create allMessages more efficient
     //List<IntWritable> allMessages = Lists.newArrayList(messages);
-    List<Long> allMessages = new ArrayList<>();
-    for (LongWritable message : messages) {
+    List<Integer> allMessages = new ArrayList<>();
+    for (IntWritable message : messages) {
       allMessages.add(message.get());
     }
     if (allMessages.isEmpty()) {
@@ -61,14 +59,15 @@ public class LabelPropagationComputation extends
    * @param allMessages All messages the current vertex has received
    * @return the maximal frequent number in all received messages
    */
-  private long getMostFrequent(Vertex<LongWritable, LongWritable,
-    NullWritable> vertex, List<Long> allMessages) {
+  private int getMostFrequent(
+    Vertex<IntWritable, IntWritable, NullWritable> vertex,
+    List<Integer> allMessages) {
     Collections.sort(allMessages);
-    long newValue;
+    int newValue;
     int currentCounter = 1;
-    long currentValue = allMessages.get(0);
+    int currentValue = allMessages.get(0);
     int maxCounter = 1;
-    long maxValue = 1;
+    int maxValue = 1;
     for (int i = 1; i < allMessages.size(); i++) {
       if (currentValue == allMessages.get(i)) {
         currentCounter++;
@@ -101,17 +100,16 @@ public class LabelPropagationComputation extends
    * @throws IOException
    */
   @Override
-  public void compute(Vertex<LongWritable, LongWritable, NullWritable> vertex,
-                      Iterable<LongWritable> messages)
-    throws IOException {
+  public void compute(Vertex<IntWritable, IntWritable, NullWritable> vertex,
+    Iterable<IntWritable> messages) throws IOException {
     if (getSuperstep() == 0) {
       sendMessageToAllEdges(vertex, vertex.getId());
     } else {
-      long currentMinValue = vertex.getValue().get();
-      long newValue = getNewValue(vertex, messages);
+      int currentMinValue = vertex.getValue().get();
+      int newValue = getNewValue(vertex, messages);
       boolean changed = currentMinValue != newValue;
       if (changed) {
-        vertex.setValue(new LongWritable(newValue));
+        vertex.setValue(new IntWritable(newValue));
         sendMessageToAllEdges(vertex, vertex.getValue());
       }
     }
